waveshare_f410

Module st7789v2

Structs

ST7789V2

ST7789V2 driver for the ST7789V2 display. This driver uses SPI for communication and requires a data/command pin, a reset pin, and a chip select pin. TODO: Implement DMA support for faster data transfer.

Enums

ColorMode 🔒

Color mode for the ST7789V2 display. This enum defines the color mode used by the display. Currently, only RGB565 (16-bit color mode) is supported.

Commands

Commands for the ST7789V2 display. This enum defines the commands used to control the display. TODO: Add more commands as needed.

Error

Error type for the ST7789V2 driver. It is a generic error type that can be used to handle errors from the SPI, CS and DC pins.