waveshare_f410::st7789v2

Struct ST7789V2

pub struct ST7789V2<'a, SPI, DC, RST, CS, const W: usize, const H: usize, const C_MODE: u8 = 85>
where
    SPI: Instance,
    DC: OutputPin,
    RST: OutputPin,
    CS: OutputPin,
{
    spi: Spi<SPI>,
    dc: DC,
    rst: RST,
    cs: CS,
    delay: &'a mut Delay,
}

ST7789V2 driver for the ST7789V2 display. This driver uses SPI for communication and requires a data/command pin, a reset pin, and a chip select pin. TODO: Implement DMA support for faster data transfer.

Fields
spi: Spi<SPI>dc: DCrst: RSTcs: CSdelay: &'a mut Delay
Implementations

impl<'a, SPI, DC, RST, CS, const W: usize, const H: usize, const C_MODE: u8> ST7789V2<'a, SPI, DC, RST, CS, W, H, C_MODE>
where SPI: Instance, DC: OutputPin, RST: OutputPin, CS: OutputPin,

pub const fn new( spi: Spi<SPI>, dc: DC, rst: RST, cs: CS, delay: &'a mut Delay, ) -> Self

Creates a new instance of the ST7789V2 driver.

Arguments

spi - The SPI interface to use for communication. must be initialized.

dc - The data/command pin, used to switch between data and command mode. when high, it is in data mode and when low, it is in command mode.

rst - The reset pin, used to reset the display.

cs - The chip select pin, used to select the display. it is active low.

delay - A mutable reference to a delay object, used for timing operations.

Returns

A new instance of the ST7789V2 driver.

pub fn init( &mut self, ) -> Result<(), Error<Error, CS::Error, DC::Error, RST::Error>>

Initializes the ST7789V2 display. This method sends the initialization commands in the order of

Software reset

Sleep out

Set color mode

Memory data access control

Display on

Returns

A result indicating success or failure of the initialization. note: that this method will block until the display is initialized. note: there is a delay after each command to allow the display to process the command.

pub fn draw_screen( &mut self, buffer: &[u8], ) -> Result<(), Error<Error, CS::Error, DC::Error, RST::Error>>

Draws the screen with the provided buffer. uses W and H constants to determine the column address and row address.

pub fn send_command( &mut self, cmd: Commands, ) -> Result<(), Error<Error, CS::Error, DC::Error, RST::Error>>

pub fn send_data( &mut self, data: &[u8], ) -> Result<(), Error<Error, CS::Error, DC::Error, RST::Error>>